// src/app/components/lit-publisher-view/interfaces/lit-publisher-view.interface.ts
export interface ILitPublisherView {
      publisherId: number;    // System.Int32
      publisherName: string | null;    // System.String
      iso3CntrRef: string | null;    // System.String
      iso2CntrRef: string | null;    // System.String
      cCountryName: string | null;    // System.String
}